Build as a universal binary for Apple Silicon
The project could not produce an arm64 slice at all. VALID_ARCHS was
pinned to "i386 x86_64", so Xcode silently skipped every compile and
link step and still reported BUILD SUCCEEDED, leaving an .app with no
executable inside it.

  - Drop VALID_ARCHS and use $(ARCHS_STANDARD), which resolves to
    arm64 + x86_64 on current toolchains.
  - Raise MACOSX_DEPLOYMENT_TARGET from 10.6 to 10.13, the oldest
    version current Xcode accepts.
  - Compile the UniversalDetector sources directly into the app
    instead of linking a prebuilt UniversalDetector.framework from
    ../UniversalDetector/build/Release. That path does not exist in a
    fresh clone, so the build failed with "the file
    UniversalDetector.framework couldn't be opened". Building the
    sources in-target matches how BGHUDAppKit and Apple Remote Control
    are already handled here, removes an embedded framework from the
    bundle, and lets the submodule inherit the app's deployment target.
  - Point the submodule URLs at https:// so a clone without SSH keys
    configured can still run "git submodule update --init".

Also casts an NSUInteger to int at a %d format argument in
SubConverter, which is undefined behaviour on 64-bit.

+++ MPlayerX/SubConverter.m
@@ -19,7 +19,7 @@
  */
 
 #import "SubConverter.h"
-#import <UniversalDetector/UniversalDetector.h>
+#import "UniversalDetector.h"
 
 NSString * const kWorkDirSubDir = @"Subs";
 
@@ -148,7 +148,7 @@ NSString * const kWorkDirSubDir = @"Subs";
 						prefix = [subPathNew stringByDeletingPathExtension];
 					}
 					// 如果该文件存在那么就寻找下一个不存在的文件名
-					subPathNew = [prefix stringByAppendingFormat:@".mpx.%d.%@", idx++, ext];
+					subPathNew = [prefix stringByAppendingFormat:@".mpx.%d.%@", (int)(idx++), ext];
 				}
 				
 				// CP949据说总会fallback到EUC_KR，这里把它回到CP949(kCFStringEncodingDOSKorean)
